Make DX's DxConsole not use static variables
Summary:
It's a pretty easy matter to make this a normal object contained in
dexer/Main.  We can use this functionality to capture output from this
step instead of spewing it to stdout.

To avoid an NPE I had to get rid of the call to makeOptionsObjects() in
Arguments.parse(), but it is always called right after parse, so there's
no change in behavior there.

Test Plan: ant dx-tests