Normalize exported and provided deps to ##ImmutableSortedSet##
Summary:
These are created and eventually stored as ##ImmutableSortedSet##
types, but the interfaces in between their creation and their
member references are generalized around the ##Set## type which
forces a copy in the ##DefaultJavaLibrary## constructor.  I'm
entirely sure if we want to standardize on ##ImmutableSortedSet##
as the main ##BuildRule## container, but it seems nice to make
this consistent and avoid the copy.

Test Plan: ant test