Support targets of the form 'android-#' for android_binary
Summary:
Pulled the regex out of AndroidPlatformTarget.getTargetForId() and modified it
to work for both 'Google Inc.:Google APIs:-#' and 'android-#' forms. The code
that finds the additional Google jars is skipped if the target is 'android-#'.
Added unit tests to test the modified regex and the modified code in
getTargetForId(). Closes #118

Test Plan:
Run `buck build hello` and verify that the build succeeds. Then replace 'Google
Inc.:Google APIs:16' with 'android-16' in fbandroid/apps/samples/hello/BUCK,
run `buck build hello`, and verify that the build succeeds.