Add object type to exopackage abi hash
Summary:
I was carefully tracking the type of each file so ensure that we would
re-run apkbuilder if a file changed type, but not path or contents.
Then I promptly forgot to add those types to the hash.

Test Plan: An upcoming diff adds a test that fails without this.