Support artifacts with executable permissions in cache
Summary:
The zip implementation we currently use doesn't perserve the
executable permission bit set on artifacts.  As a result an
executable binaries we cache (e.g. PEX files, native binaries)
become un-exectuble when being served from the cache.

This diff makes a couple changes to add support for preserving
the execute permissions on artifacts:
1) The external attribute in the central directory is used by zip
to store OS-specific permission flags (e.g. the executable bit
in POSIX).  This diff adds support for this to the `CustomZipEntry`
object and exposes this via `EntryAccounting` so that it gets
written to the zip archive when present.
2) Switch the `ProjectFileSystem.createZip` method to use the
`CustomZipEntry` class (as `ZipEntry` doesn't support external attrs)
and load the file permissions when available.
3) Switch `Unzip.extractZipFile` to use apache-commons-compress to
extract the cached zip archive, which provides access to the
external attributes.  We then detect if the user execute perms were
set in the cache, and if so set execute permissions for any roles
(i.e. owner, group, others) that had read permissions.

Test Plan: included a test