Add `MorePaths.expandHomeDir`.
Summary:
This replaces a lot of duplicated logic in `BuckConfig` and `BuckConfigTest`.

In the process of refactoring this, I "fixed" a unit test, which turned it into
a failing unit tests. Turns out there are probably a couple old unit tests we
want to delete, which I'm doing as part of this diff to avoid pushing to master
with test failures. The problem is that we used to prune absolute paths from
`getIgnoredPaths`, but when we stopped doing this the tests were left in, and
they kept passing because they were incorrectly looking for Strings in a set of
Paths.

Fixes #119, Fixes #120

Test Plan: New unit tests.