buck for iOS v2: Make Apple binary targets implement AbstractNativeBuildable
Summary:
This is a first step towards making Buck build iOS targets
natively. Building targets doesn't quite work yet (we need to transform
xcconfig settings to clang / ld parameters) but this is a start.

I introduced a new value type, TargetSources. This is necessary
because AbstractNativeBuildable requires the subclass to pass in
certain information (source and header files) to the superclass
constructor.

Since Java requires the call to super() to be the first line in
the subclass constructor, we need to make sure IosLibraryDescription
and related classes wrap up and pass the source, header, and other
required data directly to the subclass constructor.

TargetSources bundles together two existing pieces of data (srcs and
perFileFlags) with two new pieces of data (srcPaths and
headerPaths). These will be passed to the constructor of IosLibrary,
IosBinary, etc.

I also had to make the constructor to ArStep public, since it had
package privacy before.

Depends On D1386196
Depends On D1386184
Depends On D1386032
Depends On D1386009

Test Plan: Unit tests included.