Make SplitZipStep remember what files it created
Summary:
Previously, it listed the output directory to provide the list of files
to be dexed.  Now, we remember the result from the zip splitter so we
can generate them directly.

The main motivation for this is that we can use an ordered list and
generate the file names with DexStore.fileNameForSecondary based on the
index, instead of relying on specific filenames and extensions.  This is
necessary to support raw dex files, which have a different naming scheme
from our existing secondary dex files.

Test Plan: Buildbot.