Add a resources_root field to java_library and friends.
Summary:
When java_library rules have their own local resources, we usually try to put
those under one of the root-less dir names given in `src_roots` in the
.buckconfig. Sometimes that doesn't work well -- maybe the java_library dir is
already below one of the absolute paths that take precedence, or maybe the name
of the resources dir is too common to safely make it root-less. What's really
needed is for a java_library to be able to specify a resources directory that
only applies to that one rule. Thus, the new `resources_root` field.

Test Plan: New unit and integration tests.