Add resources_root to the rule key of DefaultJavaLibrary.
Summary:
I forgot to do this when I defined the resources_root field. This
makes sure the rule is rebuilt if the resources_root is changed.

Test Plan:
Define a random library in fbandroid, change the resources_root field, and
confirm that it rebuilds. (Also that it didn't before.) Note that this doesn't
repro in the buck project (that is, changing resources_root does cause a
rebuild, even without this diff), and even setting cache.mode doesn't make it
repro.  What config could it be?