Buck iOS SDK: Read default compiler from SDKSettings.plist
Summary:
The next step in building iOS code with Buck is to
apply the correct compiler and linker flags as defined by
.xcconfig files.

The meaning of the key-value pairs in .xcconfig files is
defined in .xcspec files bundled in Xcode plugins.

To discover which .xcspec file to use, we need to know which default
compiler Xcode wants to use. This (along with other default xcconfig
settings) is defined in a file SDKSettings.plist that comes with
Xcode.

This diff adds a new class com.facebook.apple.xcode.xcspec.SDKSettings
which parses the SDKSettings.plist file and extracts the default
settings (including the default compiler).

Test Plan: Integration test included.