Clear interrupted state after WatchmanWatcherTest
Summary:
This was being set and lingering around, breaking subsequent tests.
This was not a problem for "buck test" because it uses a separate
process for each test module, but "ant java-test" runs them all in the
same process.

Test Plan:
MoreFuturesTest passes.  I also added some logging to the start of it to
print the interrupted flag.  Without this diff it was true.  With it,
false.