Amend AndroidPackageableCollector.addResourceDirectory() to take a Supplier for the package.
Summary:
This is needed for upcoming work with .aar files.

Specifically, at the point when `AndroidPackageable#addToCollector()` is invoked,
the `rDotJavaPackage` for an `.aar` is not yet known because it must be unzipped
and have its `AndroidManifest.xml` parsed to figure it out. All of this will be
done before the `UberRDotJava` rule is built, so it should be fine, in practice.

Test Plan: Sandcastle builds.