Buck logging: Always rotate logs on daemonized command
Summary:
In D1413903, I added LogConfigFilesWatcher to reload
logging config (and, as a side effect, rotate logs) when logging
config files changed.

That's not actually the behavior we want: we want to unconditionally
rotate logs each time we start a command while daemonized. The only
documented way to do that is by reloading config, so this diff
makes it so we always reload config when starting a command while
daemonized.

Test Plan:
In fbandroid, ran buck build fb4a twice. Before, logs were
not rotated. Now, they are.