If package is not specified for an android_resource(), try to extract it from the manifest.
Summary:
Historically, we provided the option to specify the package for `R.java` via the `package`
argument to `android_resource()`:

```
android_resource(
  name = 'example',
  package = 'com.example',
  res = 'res',
)
```

This eliminates the burden of creating dummy `AndroidManifest.xml` files of the form:

```
<manifest xmlns:android='http://schemas.android.com/apk/res/android' package='com.example' />
```

However, when importing resources from an `.aar`, it would be equally burdensome to require
users to do something like:

```
android_prebuilt_aar(
  name = 'example',
  aar = 'example.aar',
  package = 'com.example',
)
```

as the `package` argument duplicates information that is already present in the `AndroidManifest.xml`
that is in `example.aar`.

As such, we introduce the following new behavior:

* If there is no `res` argument, then no attempt is made to determine the package.
* If there is a `res` argument and `package` is specified, its value is used.
* Failing that, if there is a `manifest` argument, the `package` is parsed from it.
* If there is a `res` but no `package` or `manifest`, an exception is thrown.

An important side effect of this change is that `AndroidResource` now passes `rDotJavaPackage` as a
`Supplier` to `AndroidPackageableCollector`, which is needed for upcoming work with `.aar` files.

Test Plan:
Added test cases to `AndroidResourceTest.java`.
Added `AndroidResourceParsePackageFromManifestIntegrationTest`.
Sandcastle builds.