Break GenRDotJavaStep into two steps: generate manifest; run aapt.
Summary:
Instead of generating the `AndroidManifest.xml` under a random tmp directory,
we now write it in a predictable location under `buck-out`.

Previously, some I/O was done in the constructor of `GenRDotJavaStep`,
which was not appropriate. This change also makes it possible to eliminate a
hack in the superclass, `ShellStep`, that was added to support this bad
behavior.

Test Plan: Sandcastle builds.