Use better names for enhanced JavaLibrary rules for BuildConfig.class files to add to an APK.
Summary:
When doing graph enhancement in `AndroidBinaryGraphEnhancer` for build configs,
instead of adding a `JavaLibrary` to `AndroidPackageableCollection.classpathEntriesToDex`
via `AndroidPackageableCollector.addClasspathEntry()`, we now create a `JavaLibrary` and
a `DexProducedFromJavaLibrary` explicitly so we can control the names of these enhanced
rules and ensure they are sane. The `DexProducedFromJavaLibrary` is added iff pre-dexing
is enabled.

An important thing to note is that now `AndroidPackageableCollection.classpathEntriesToDex`
does not have all of the classpath entries that the `AndroidBinary` needs to dex.
Instead, `EnhancementResult.getClasspathEntriesToDex()` has the complete collection.

This also fixes https://github.com/facebook/buck/issues/187.

Test Plan:
`MultipleBuildConfigIntegrationTest`
Sandcastle builds.