Introduce android_prebuilt_aar() rule.
Summary:
An `.aar` file is a zip file that serves as the binary distribution of an Android Library Project,
as specified by http://tools.android.com/tech-docs/new-build-system/aar-format.

This diff makes it possible to depend on an `.aar` file as if it were an `android_library()` rule:

```
android_prebuilt_aar(
  name = 'play-services',
  aar = 'play-services-4.0.30.aar',
)

android_library(
  name = 'lib',
  # This Java code can compile against Play services and reference its resources.
  srcs = glob(['*.java']),
  deps = [ ':play-services' ],
)
```

Note that the `aar` argument to `android_prebuilt_aar()` could also be the output of a `genrule()`,
so now it is effectively possible to generate an `android_resource()` rule. Specifically, if you
generate the contents of a `res` folder and properly package it into an `.aar` zip file (which is the
output of a `genrule()`), then you could specify that `genrule()` as the `aar` argument to an
`android_prebuilt_aar()` so others can depend on your generated Android resources. You can
see an example of this in `AndroidPrebuiltAarIntegrationTest`.

This diff addresses the first half of https://github.com/facebook/buck/issues/117.

Test Plan:
`AndroidPrebuiltAarIntegrationTest` verifies that `android_prebuilt_aar()` can take a genfile
that conforms to the `.aar` package specification and serves as an Android library project
that another `android_library()` rule can build against.

In `test/com/facebook/buck/android/testdata/android_prebuilt_aar`, I ran:

```
touch .buckconfig
buck install //:app --run
```

and verified that the app started successfully.