Convert the manifestFile param from Path to SourcePath.
Summary:
Ultimately, this was fueled by the need to change
`AndroidPrebuiltAarGraphEnhancer` to pass the manifest as an `OutputOnlyBuildRule`
because it was being returned by `AbstractBuildRule#getInputsToCompareToOutput()`
for `AndroidResource` when it should not have. This was causing a
`FileNotFoundException` in `DefaultFileHashCache`.

Test Plan:
Run `rm -rf buck-out buck-cache` and build an app that uses
`android_prebuilt_aar()` and verify that `DefaultFileHashCache`
no longer throws a `FileNotFoundException` where it did before.