Introduce OutputOnlyBuildRule and delete InternalSourcePath.
Summary:
`InternalSourcePath` was originally introduced to deal with the case
where a `Path` under `buck-out` needed to be referenced as a `SourcePath`,
but the `Path` was not the output of any particular `BuildRule`, so a
`BuildRuleSourcePath` could not be used.

This diff changes things so that we use a synthetic build rule,
`OutputOnlyBuildRule`, to wrap such a `Path` so that a `BuildRuleSourcePath`
can be used. This introduces fewer new concepts into Buck's execution model.

Test Plan: Sandcastle builds.