Make BuildTarget.getBasePath() return a Path
Summary:
There are several places where we're using Path.getBasePath and coercing it into a Path.
We also use the output of that method to ask a BuildTarget whether it's in the root of
the project. Avoid both of those cases by making getBasePath return an actual Path.

Test Plan: buck test --all