More verbose Assumption Violations
Summary:
Print the stack-trace of an assumption violation, so we can see what failed.

We're using assumption violations to mean "something went wrong, but it doesn't mean the test is failing".  Accordingly, we always want to see what actually went wrong.  The test-console shows the stack trace, so we should also do that at the command line.

It's possible (as per the original task) that this should be configurable -- perhaps some people expect assumption violations to be less noisy, causing tests to be skipped because they couldn't be run for some benign reason, as opposed to because something non-test related failed badly?

Test Plan: Run it with the test buck project that has assumption violations.