Add option project.default_exclude_paths
Summary:
This adds the ability for a repo to exclude one or more
paths from processing in 'buck project'.

This is useful for repositories that have sandbox directories
that hold Buck projects which may or may not build at any
given time. (Unlike 'project.ignore', an excluded path can
still be built if specified explicitly).

Currently, I only implemented it for Xcode projects, but I'll
file a followup task to ensure it's implemented for IntelliJ
projects as well.

Test Plan:
Unit test included. Ran on repository with a
directory which I added to project.default_exclude_paths.
Confirmed in logs that directory was excluded when I ran
'buck project'.