Add ability to use an HTTP cache
Summary:
Now you can specify "http" as one of the cache modes in a
buckconfig file, and Buck will communicate with the specified host via
HTTP on port 5551. Docs are updated to describe the new options.

Test Plan:
Run a basic http server that listens on port 5551. Update the
.buckconfig file to use "http" as the cache mode. Build a project and
verify that no error messages are printed by Buck. Also verify that the
tests pass.