Test script for the exopackage installer.
Summary:
An upcoming diff will make some changes to the installer to improve
device compatibility.  This is the script that I used to test it.  It's
not an exhaustive test of exopackage (for example, it doesn't verify
that we don't re-install the apk on a Java-only change), but it does
test the happy path of doing a clean and incremental install and verify
that the code changes.

There's a very small app in testdata that logs a string on startup.  The
script builds it, installs it, captures logs, edits it, builds it,
installs it, captures logs, then verifies that both logs are correct.

There's a wrapper script that runs the test serially on all connected
devices, then prints a result summary.

Test Plan:
Ran it on a bunch of devices.  Saw mostly successes, but some failures,
which were properly reported.