Update android_build_config() to support custom values.
Summary:
Previously, `android_build_config()` generated a `BuildConfig`
file with only two constants: `DEBUG` and `IS_EXOPACKAGE`.
This diff introduces a feature that enables developers to include
additional, custom fields to `BuildConfig`. Like `DEBUG` and `IS_EXOPACKAGE`,
the values of these fields will not be constant expressions at compile
time, but can be constant expressions before ProGuard is run when
packaging the `BuildConfig.class` into an APK.

Specifically, this diff adds two optional arguments to `android_build_config()`, `values`
and `values_file`. These arguments can be used to specify additional fields (and values)
for the generated `BuildConfig` class. The `values` argument should be used when
the values are static and can be specified directly in the build file. In contrast, the
`values_file` argument should be used when the values are determined by the output
of some other build rule.

If specified, `values` is a list of strings that match the pattern:

  <JavaType> <JavaVariableName> = <LiteralValue>

For example, the following would be a valid declaration of `values` in a build file:

```
android_build_config(
  name = 'build_config',
  package = 'com.example.buck',
  values = [
    'int VERSION = 1234',
    'boolean IS_A_GOOD_FIELD = true',
    'String APP_ID = "vcebelutvgnvljrnnvuurudbbknrbfij"',
  ],
)
```

If the `values_file` is specified, it must identify a file whose contents are lines
like those in the `values` argument in a build file. For example, valid content of
a `values_file` would be:

```
int VERSION = 1234
boolean IS_A_GOOD_FIELD = true
String APP_ID = "vcebelutvgnvljrnnvuurudbbknrbfij"
```

If both `values` and `values_file` declare fields with the same name, the values
in `values_file` take precedence.

To override any of the values for the fields in an `android_binary()` that transitively
depends on the `android_build_config()`, specify at least one of the `build_config_values` or
`build_config_values_file` arguments.
Like `values`, `build_config_values` must specify a string array while
`build_config_values_file` must specify a file whose contents match the field declaration
format. Not all fields must be redefined to use this feature.

Note that both `values_file` and `build_config_values_file` are `SourcePath`s so they can
either be static files in version control, or files generated from another build rule.

This diff also updates the public documentation for `android_build_config()`
to explain this new feature.

Test Plan: Updated corresponding unit and integration tests.