Extract both target GIDs and target names in ProjectParser.
Summary:
To generate Xcode schemes that xctool can read, we need
four things for each target:

1) The target's GID ("BlueprintIdentifier")
2) The target's Xcode project ("ReferencedContainer")
3) The target's name ("BlueprintName")
4) The result of building the target ("BuildableName")

We were previously missing 3) and 4), but in D1439254, I added them by
reading the PBXFileReference for each target.

This worked well for everything except xcode_native targets, which use
a fake target (see
WorkspaceAndProjectGenerator.generateWorkspaceAndDependentProjects()).
This fake target doesn't have a PBXFileReference,
so my logic in D1439254 crashes with an NPE.

We can get the data we need for xcode_native targets by building on
the existing ProjectParser call added in D1311030 to extract not
just the (target -> GID) maps, but also the (target -> filename)
maps for the BuildableName in the scheme.

This adds a new API
ProjectParser.extractTargetNameToGIDAndFileNameMaps() building on the
old API ProjectParser.extractTargetNameToGIDMap(), along with tests.

Test Plan:
Unit tests included.