Ensure generated schemes correctly handle xcode_native targets.
Summary:
This is a follow-up to use the new ProjectParser API added
in D1459445 to fix the NPE Buck throws when generating schemes that
include xcode_native targets.

Test Plan:
Ran buck project --workspace-and-projects in fbobjc.
Before, NPE. Now, it correctly generates the workspaces.

There are no tests yet for WorkspaceAndProjectGenerator;
I filed a task to address this.