Allow specifying GID in iOS target settings.
Summary:
Currently, there are two ways to set the Xcode "global ID" (aka "gid")
for a target:

1) Automatically generated using a stable hash algorithm
2) From an existing project file on disk

together.

in to source control.

During a transition, some repositories will be in a blended state where
they check in non-generated workspaces and schemes which refer to generated
projects. Those schemes will contain GIDs referring to projects on disk.

To make the transition easier, we'll allow a new method to set the Xcode GID
for a target:

3) Specified in the "gid" parameter in the BUCK file

This will allow us to generate the projects with a specific GID so the
existing schemes can continue to work.

Test Plan: Unit tests updated.