Introduce a TargetGraph class so that BuildEngines can make use of it.
Summary:
There are commands, such as when auditing owners, that only need access
to the TargetGraph and not the ActionGraph. Expose this as a type that
can be accessed from the Parser.

A later diff will move construction of the ActionGraph out of Parser
and (probably) into this class.

Test Plan: buck test --all