Pass environment to daemon methods.
Summary:
Once the daemon can be used by multiple commands concurrently
it will no longer be able to rely on always being able to use a single
environment as every method call may come from a different, concurrently
executing, client. This change anticipates concurrent multiple
commands by removing the environment from the Parser. If the Parser
is asked to parse a build file with an environment that differs from
the one used to populate the cache, the cache is invalidated and the
build files reevaluated. This mimics the behaviour used to invalidate
the Parser cache when the default include files are changed.

Test Plan: buck test --all