Create the RepositoryFactory.
Summary:
In order to support cross repo targets properly, we need a Repository to be
able to reference other Repositories, with in turn reference others, and so on.
We don't want to create all these up front, both because we want to avoid the
IO, and because we don't want to throw errors on missing repositories unless
the user actually needs them. So we need to abstract away repo creation into
a class that I'm calling `RepositoryFactory`, because Java has poisoned me.

Test Plan:
Unit tests still pass. Many of them now rely on the new
`TestRepositoryBuilder`.