Use Jacoco to create a code coverage report; update build scripts to run it.
Summary:
We used to measure code coverage with EMMA, but it proved to be too flaky to
run on every build. This diff introduces a script to measure it with Jacoco:

```
buck test --all --code-coverage --code-coverage-format xml --no-results-cache
python scripts/assert_code_coverage.py
```

The script prints out a summary of code coverage by package to the console.
It returns with a non-zero exit code if code coverage by line drops below 78%.
Currently, we are at 81.06%, so hopefully we can maintain this (or even better,
crank it up over time).