Fix http cache timeout units.
Summary:
Previously the code interpreted the config value as milliseconds
when it was actually supposed to be seconds. I also introduced a new
config value http_timeout_seconds separate from the existing timeout for
cassandra, which is more consistent with the other http_* options.

Test Plan:
Built some apps with buck, verified nothing fails with
http_timeout_seconds set to a low number.