Only serialize test result events to a file.
Summary:
Writing all events to a file is overkill. Many large projects generate
huge numbers of events, and existing use-cases of
`buck test --output-events-to-file ...` are only concerned with test results.

This diff reduces the number of event subscriptions down to the bare
minimum -- the event containing test results.

Test Plan:
Patched buck on my dev server and checked that the test-related events were
still written to a file.