Reuse ObjectMapper in WatchmanWatcher.
Summary:
WatchmanWatcher always re-creates a JsonFactory to do its
JSON parsing. This is not good for perf:

http://wiki.fasterxml.com/JacksonBestPracticesPerformance

This diff builds on 7569550 to re-use JsonFactory via a passed-in
ObjectMapper. I plan to build further on this to use the ObjectMapper
for JSON formatting of the query in a subsequent diff.

Test Plan:
```
ant test -Dtest.class=WatchmanWatcherTest
ant test -Dtest.class=WatchmanIntegrationTest
```