Remove assumption project and test rules must be defined in the same BUCK files as the rules they relate to.
Summary:
The project command makes the assumption rules that relate to other rules in ways that aren't deps
(test rules' sourceUnderTest and project rules' projectRule) are always defined in the same BUCK file as the
rules they relate to. There is no such limitation in practice however, and it can cause the project generator
to miss them if they're not.

Test Plan: new tests