Don't resolve the base path before passing it to the project generator.
Summary:
ProjectGenerator expects to be passed relative paths, but ever since RepositoryFactory
was introduced projectFilesystem is constructed with an absolute path, so resolving
basePath against it will return an absolute path.

As base path is already equal to the relative path from the repo root to the output directory
there should not be any need to resolve it.

Test Plan: CI, try to generate a workspace that has projects that depend on asset catalogs.