Javadoc classpath should be limited to src/ directory.
Summary:
This reduces the classpath for the Javadoc to the appropriate scope.
Because we do not publish the Javadoc for our tests, this also replaces
one `{@link}` with `{@code}` where we referenced a test.

Test Plan:
Changing the classpath triggered the extra warning, which indicated we
changed it to the right thing. Fixed the warning in the .java file and
verified it is no longer reported by `ant javadoc`.