Mini aapt - step to compute R.java declarations from a set of android resources.
Summary:
This diff introduces a new step which will be used instead of `aapt` to generate
the `R.txt` file for an `android_resource` target. I'll send out subsequent
diff that actually makes `AndroidResource` use this step.

In addition to generating `R.txt`, the new step also checks for references in a
particular resource module against a set of dependencies  (which will be its
first order dependencies, in the subsequent diff).

Test Plan: Unit tests.