Only include the main target's dependencies in the scheme.
Summary:
When generating workspaces, the generated scheme would include all targets
of all the projects required to build the main target and its dependencies. This
changes it to only include the strictly required targets. Note that this means
tests are not included in the scheme anymore unless they're depended upon or main
targets, but they were only being included as a side effect of being in the project
originally, so I don't consider that a regression.

Test Plan: Generate a workspace for an app that includes tests and asset catalogs.