Ensure xcode_native dependencies are linked in to binary
Summary:
To refer to libraries built by existing Xcode projects, we
use the 'xcode_native' rule type. It correctly expresses the
dependencies, but binary and test rules which depended on xcode_native
rules were not actually linking the library into the final result,
which meant we had to hard-code linking them in via the BUCK file.

This fixes the bug and adds a test.

Test Plan: Added test. It failed. Applied fix, test no longer failed.