Make test coverage report with JaCoCo work in the wild
Summary:
EMMA coverage report invocation is working, but it is somehow flaky.
Test coverage report with JaCoCo is looking very promising but it
seems to only work for Buck project itself because some paths are
hard coded and other paths are customizable but still not exposed
in buck_common script as it's the case with broken EMMA.

Make hard coded path customizable and expose the paths in buck_common
so that folks in the wild can use it.

Closes https://github.com/facebook/buck/pull/173

Test Plan:
Clone Gerrit Code Review project and issue this command:
buck test --all --code-coverage --code-coverage-format html --no-results-cache
Enjoy the report in buck-out/gen/jacoco/code-coverage/index.html