Add the ability to configure libthrift for JavaThriftLibrary in .buckConfig
Summary: Generated java thrift files depend on libthrift (which also
depends on slf4j as an exported dep).  Add the ability to specify the
libthrift to be used for the java flavor of the thrift library so that
source can be compiled.

Test Plan: Test runs, use it to compile java thrift_libaries