Add support to XcodeNative for custom target and buildable names
Summary:
Previously, XcodeNative assumed that the target name
was the same as the rule name, and the buildable name was always
"lib[rulename].a".

This isn't true for all native Xcode projects, so this diff
adds the ability to customize the target and buildable name.

Test Plan:
Unit test included for project generation.
Workspace generation unit tests pending #4783327; tested
by running 'buck project --workspace-and-projects //path/to/target:workspace'
for a workspace containing references to xcode_native targets
with custom library names.