Allow the classloader used by annotation processors to be garbage collected early.
Summary:
There are reports of large, complex builds using lots of annotation processors
running out of PermGen space on Java 7. The incidence of this increased markedly
after the introduction of the classloader used for loading annotation processors.
I suspect that _something_ is holding a reference to the Step which, in turn, is
holding one to the classloader, which could get pretty big.

As an interim measure, we'll null out the classloader after compilation is
finished. My experiments suggests that this remedies the problem. For now....

Test Plan:
Run a large build that exhausts PermGen without this patch
Watch it fail
Apply diff and run the build
Gasp in amazement as it passes