Fix minor bug in missing symbol finder
Summary:
The missing java symbol support, which hunts down rules that resolve
missing symbol errors in java, parses BUCK up the dir tree in an effort
to find the owners of sources.  It determines that a file is a BUCK file
by checking that it's basename is BUCK.  However, it's possible that
the user has a directory by this name, and this is even more likely
on case-insensitive filesystems where a directory called "buck" would
be mistaken for a BUCK file and parsed, resulting in an ugly parse
error.

This diff changes the code to check that the given path is a file
to avoid this.

Test Plan:
Removed some needed deps in test code and ran ##buck test --all##.