Copy phase refactor: Introduce ProjectGenerator.addCopyFilesBuildPhase()
Summary:
ProjectGenerator.addCopyFrameworksBuildPhase() is the only step in the
current project generator that adds a PBXCopyFilesBuildPhase to the
project.

However, we'll need to add a PBXCopyFilesBuildPhase to copy application
extensions into the bundle's PlugIns directory. This diff extracts out
the logic to add a PBXCopyFilesBuildPhase to the project into a new
routine, addCopyFilesBuildPhase(), so we can use it in later diffs to
support extensions.