Load the AbiWriter from the current classloader.
Summary:
We're still seeing memory leaks from the java compilation. Most of the culprits point to
the AbiWriter not being properly garbabge collected. We know that the AbiWriter only uses
classes from the JRE, so it's fine for us to load it from the current classloader without
worrying it might accidentally pollute the classpath of classes being compiled.

With this fix in place, compilation of a large java codebase keeps PermGen usage to
around 30-40MB, and usage appears to return to normal levels at the end of a build.

Test Plan:
buck test --all --code-coverage
Check that coverage looks good
Build a large codebase and watch PermGen usage in jvisualvm