[iOS header maps] Fix typo in build setting for user includes
Summary:
[iOS header maps] Fix typo in build setting for user includes

Apparently I shot myself in the foot with a search/replace in an early version of the diff.

For consistency of the vocabulary, I also replaced flat->user everywhere with Intellij.

Apple's documentation of USER_HEADER_SEARCH_PATHS is here:
https://developer.apple.com/library/mac/documentation/DeveloperTools/Reference/XcodeBuildSettingRef/1-Build_Setting_Reference/build_setting_ref.html#//apple_ref/doc/uid/TP40003931-CH3-SW21

Test Plan: unit test + verify that the build setting works.