[iOS] Do not add both classic and header-map search paths for the same libraries
Summary:
[iOS] Do not add both classic and header-map search paths for the same libraries

This did not seem to be a big deal first, but long compilation command lines tend to slow down the build.

Incidentally, I noticed that buck-out is never flushed, so including header maps that were not generated this time is not correct.

Note: When we have more projects in 'buck header map mode', we should also consider Yiding's suggestion to compute the transitive closure of dependencies in advance in the header maps themselves, in order to reduce the number of lookups (at the price of some more disk space and arc-build time).

Test Plan: unit tests