Pass ordered build rules into SchemeGenerator
Summary:
SchemeGenerator currently calculates the ordered set
of build rules internally.

I need that set in WorkspaceAndProjectGenerator
(to determine which Xcode projects to include in the generated workspace),
so this diff pulls the calculation of the set of build rules up into
that class and passes it down into SchemeGnerator.

I moved the logic into AppleBuildRules since we also need it in tests.

Now that we're passing the ordered set of build rules into
SchemeGenerator, we no longer need the hacks which worked around the
fact that the topological sort wasn't stable, so I removed them.

Test Plan: ant test -Dtest.class=SchemeGeneratorTest