Use mini-aapt to generate R.txt of android resources.
Summary:
Previously, we were invoking `aapt` to generate the R.txt file for an
android_resource() rule, and passing it resource directories from all the
transitive android_resource depenedencies.

With this diff, we use `MiniAaptStep` to generate a similar `R.txt` file,
but it only contains symbols from the android resource module and not its
transitive dependencies. Also, we only pass the first order dependencies
of the android resource rule. As a result, the ABI key also includes just
the first order deps now.