Add `extendFlavoredBuildTarget` method
Summary:
This is similar to `createFlavoredBuildTarget` but accepts build
targets which are already flavored, in which case it creates and
uses a new flavor formed by concatenating the original flavor and
the new flavor.  This is useful for flavored target nodes (e.g.
C++ thrift support, C++ debug/optimized build mode support) which
uses flavors to name the potentially many build rules it creates
(e.g //foo:bar#cpp-libbar.a).

Test Plan:
- included tests
- used in and thrift C++ support to name build rules