Allow python rules to depend on C++ libraries
Summary:
This diff makes C++ libraries a python packageable, allowing a
python binary/library to depend on C++ library rules.  This types of
deps cause the python package to pull in the shared library version
of the C++ library and map it into the python package by it's soname.

This doesn't actually add shared lib support to the python executable
archive format (e.g PEX), but just gets the buck-side prereqs out of
the way.

Test Plan:
- included tests
- built PARs in fbcode by overriding path_to_pex in .buckconfig