New class CommandThreadAssociation factored out of LogFormatter
Summary:
Currently, the LogFormatter class contains a helper
class used to manipulate the <thread ID: command ID> map.

Since I need to factor that map out to fix #4985049, it made
sense to factor out the static classes as well; they really
don't belong there any more.

This factors out the CommandThreadAssociation class from LogFormatter
into a new class. A subsequent diff will make use of it.

I removed the Closeable implementation as well, since @natthu and
I discovered that using a try-with-resources block for this is
inappropriate, as we need to log inside a catch block, which depends
on the <thread ID: commmand ID> map.

Depends On D1519662

Test Plan: Tested in subsequent diff.