Rename paired events to related events.
Summary:
"Paired" events is a unintuitive name. Every event is paired
with itself, and has to be so that the AbstractBuckEvent#equals(Object)
implementation works. The comment to BuildEvent#eventsArePair(BuckEvent)
suggest using it to pair up start and finished events, but the method
returns true for two start events or two finished events, when they
obviously don't pair up.

Test Plan: CI