Stop storing request bodies on HttpResponse objects, to avoid OOM.
Summary:
We were hitting OOM errors in particularly large builds. The
stacktrace pointed to our build listener, on the line where it serialized build
events into a network request for logging.  These requests use a
BlockingHttpEndpoint to avoid queueing up too many requests in memory for
exactly this reason. Unfortunately, the Futures for these requests are still
accumulated in a big list, and those were storing references to the request
bodies. Get rid of the references.

Test Plan:
Artificially inflate the size of request bodies to repro the OOM
error. No longer works after this diff.