Migrate one test from the root testdata directory into its proper place.
Summary:
This moves `testdata/javatestrule` to
`test/com/facebook/buck/java/testdata/java_test_get_class_names`.
The root `testdata` directory is a historical artifact for integration
testing that predates `ProjectWorkspace`.

Because only some of the tests in `JavaTestRuleTest.java` need
`ProjectWorkspace`/`DebuggableTemporaryFolder`, I split out the
ones that needed it into a separate test,
`JavaTestGetClassNamesIntegrationTest.java`. This makes it easier
to apply a common `setUp()` method to all of the tests that
actually touch the filesystem.

Test Plan: Sandcastle.