Ensure schemes do not include each others' tests
Summary:
This diff ensures that we only pass in to SchemeGenerator the
test-building and test-running rules for rules included in
the recursive dependencies of the main rule in the scheme.

I added a new integration test
`ProjectIntegrationTest.testBuckProjectGeneratedSchemesDoNotIncludeOtherTests`
which includes two schemes with overlapping but disjoint sets of tests.

Part of this fix slightly adjusted the order in which we run tests (we run
dependencies before the main rule now), so I had to tweak one of the
other tests.

Test Plan: Integration test included.