Support source maps for python sources.
Summary:
Allows users to specify their input sources as either a list of
sources or a map of base-path relative module names to input sources.
In the former case the module name to input source mapping is inferred
(as is currently done now).  In the latter case, the user can specify
an explicit module layout.

Test Plan: Included a test for the functionality in `PythonUtil.toModuleMap`