Made some PBXTarget fields non-nullable.
Summary:
PBXTarget has several nullable fields as it's designed
to be used as a mutable value and configured after creation.
Unfortunately that makes static analysis harder.
This changes some of those fields to be non-nullable, as
their value was always available at construction time or
was a mutable value type and was never changed.

Test Plan: CI