Plumb a RepositoryFactory down to the Parser.
Summary:
We will need this to actually parse cross-repo targets. Most of the
lines changed in this diff are just updating constructors in tests.

As part of doing this, I've wrapped a number of Exceptions in RuntimeException,
so that I don't have to update all the signatures in Parser and all of its
callers with checked exceptions. I'm not sure which is actually worse. Please
advise.

Test Plan: Lots of tests affected, still passing.