Remove CxxHeader rule.
Summary:
The CxxHeader rules were used to track changes to a C/C++ descriptions
header files, and a dependents that transitively depended on that
description would have it's compile rules depend on the CxxHeader rule
so that header file changes would trigger recompiles.  This diff removes
the middleman and just has each compile rule grab a full map of it's
transitive includes, allowing a future change where the compile rule
can fixup it's error output so that we show the real header file paths
that back the symlink tree view we setup.

Test Plan: buck test --all