Easy: Remove main target graph from WorkspaceAndProjectGenerator
Summary:
I'm adding support for multiple test targets in a workspace.

When doing so, I noticed the main target graph was totally unused in
WorkspaceAndProjectGenerator; we only used it to do a (target -> rule)
lookup in one location.

This diff removes it and passes in the rule directly.

Test Plan: ant test -Dtest.class=ProjectIntegrationTest