Make SchemeGenerator primary rule optional
Summary:
We'd like to make Xcode workspaces which have schemes that
contain no Launch or Profile actions, but only a set of Test actions.

This makes SchemeGenerator's primary rule (used for the Launch and Profile
actions) optional, and adds a test to confirm the desired behavior when
the primary rule is not present.

Test Plan: ant test -Dtest.class=SchemeGeneratorTest